Return actually reachable IP as part of the Floating IP API
The Floating-IP list API is only ever queried for getting the IP assigned
to a specific port (if vim-emu is added with the manual floating-ip
assignment turned off). Thus, the previous solution lead to fixed IPs being
returned (which are not reachable).

This solution tracks the assigned container for each port in order to be
able to return the public IP assigned by docker.

This is not perfect in the sense that multiple ports do not receive
different floating ips, but at least the containers are actually reachable
now.

Change-Id: I233e5de35b3b7e7350faaef017711f7c1e919fff
Signed-off-by: schillinge <ablu@mail.uni-paderborn.de>