Support (simple) classifiers
In order to support classifiers, each flow entry needs to specify the
classifier's parameters within its match rule. This means that the
created rules do not only depend on the ingress and egress port, but
also on the classifier (this required to move the creation code into the
flow classifier loop).

Change-Id: I96817255580462e0ffb570b025b6f912f2d00b13
Signed-off-by: schillinge <ablu@mail.uni-paderborn.de>