Provide API for allowing full stack emulation
This commit adds an API in order to emulate the full OSM stack. Different
levels of API are added in order to allow flexible scenarios.

The lowest level API only wraps the required components in containers.
The next higher level also abstracts the composition of these low-level
into a working OSM instance.
The third level also abstracts the emulation and configuration of the VIM
layer.

Examples are added for each abstraction level.

Implements Feature 7291.

Change-Id: I156f6625d4ff4010d688c41419c4ece03a339937
Signed-off-by: schillinge <ablu@mail.uni-paderborn.de>